Make it possible to change the final reducer for a value defintion.
ValueDefinitions all operate on arrays of numbers, but the final value for an edge type for a node is a
single number. A reducer takes an array of numbers and returns an array with a single number. When all is
said and done the final reducer is computed and then the number in the final array is used.

Currently only two reducers exist: 'first' and 'mean' (default).

Part of #3.