Make it so when calculating values we keep a partial result
And set default valeues only at the end.

This means that in the next commit when we add ValueDefintionResultValue and pass partialResult, the values
that aren't updated yet won't be passed to the calculator, making it less likely there will later be a bug
of relying on values that aren't listed in dependencies.

Part of #3.