Fix broken/odd behavior in how node value overrides worked.
Before, if you had a node override set then it would override everything else, even if a scenario was suppsoed to override it.

And a scenario override for that node that used ValueDefinitionInput would get the value from ROOT to base on, not the actual value it owuld have been had the scenario not applied.

this was out of a mis-guided attempt to save computation costs in override situations. But it's not valid if you can do full ValueDefinitionInputs for scenarios (and soon, for node overrides too).

The flow is now: 1) root, 2) base calculation, 3) node override, 4) scenario override.

This sets us up to allow phase 3 to also modify the base calculation output, like scenario override does.

Part of #3.