Make it so when you begin editing a scenario node value, it is set to the current value to start.
Before, the UI behaved mostly correctly, but as a (helpful) bug, because the value in the scenario overlay was set to 0, but the value in the UI was set to the previous value. Then, on the first
update, tjhe overlay would be set correctly.

Now, we actually set the value in the overlay explicitly.

It's kind of weird to have to pipe through the starter value, instaed of being able to just compute it in the action creator, but it's non trivial to fetch the right one so whatever.

Part of #24.