Make the default stroke for edges white.
The default width is still 0. This means that if there is a stroke, it will default to being visible. Before
with the same color as the fill of the node, it defaulted to making it look like hte radius was larger,
creating weird styling including making it look like the radius was changing.

Part of #22. Part of #24.