Factor out renderEdges off of node and make it take an array of nodes.
Also factor out spreadBumps and edgeDefinitionHelper.

Technically the way it handles multiple nodes in renderEdges (if it were ever passed it) is wrong because some of the inputs to the ValueDefinition would be tied only to the first node, not any
subsequent nodes... but also it's currently only called from a single node context.

No difference in result (yet).

Part of #12.